+++ io/esound/src/esoundpmo.cpp
@@ -392,7 +392,7 @@ void EsounDPMO::WorkerThread(void)
       }
 
       iRet = write(audio_fd, pBuffer, m_iDataSize);
-      if (iRet < 0) 
+      if ((int)iRet < 0) 
       {
           m_pInputBuffer->EndRead(0);
           ReportError("Could not write sound data to EsounD.");
